+++ grails-app/services/org/pih/warehouse/importer/LocationImportDataService.groovy
@@ -7,32 +7,22 @@
  * the terms of this license.
  * You must not remove this notice, or any other, from this software.
  **/
-package org.pih.warehouse.data
+package org.pih.warehouse.importer
 
 import grails.gorm.transactions.Transactional
 import grails.validation.ValidationException
-import org.pih.warehouse.core.Address
+import org.pih.warehouse.LocalizationUtil
 import org.pih.warehouse.core.Location
 import org.pih.warehouse.core.LocationGroup
+import org.pih.warehouse.core.LocationService
 import org.pih.warehouse.core.LocationType
-import org.pih.warehouse.core.LocationTypeCode
 import org.pih.warehouse.core.Organization
-import org.pih.warehouse.importer.ImportDataCommand
-import org.pih.warehouse.inventory.Inventory
-import org.pih.warehouse.LocalizationUtil
-
-import javax.annotation.Nullable
 
 @Transactional
-class LocationDataService {
+class LocationImportDataService implements ImportDataService {
+    LocationService locationService
 
-    def organizationService
-    def identifierService
-
-    /**
-     * Validate inventory levels
-     */
-    Boolean validateData(ImportDataCommand command) {
+    void validateData(ImportDataCommand command) {
 
         def locationNamesToImport = command.data.collect {it.name}
         def locationNumbersToImport = command.data.collect {it.locationNumber}
@@ -119,89 +109,10 @@ class LocationDataService {
 
     void importData(ImportDataCommand command) {
         command.data.eachWithIndex { params, index ->
-            Location location = createOrUpdateLocation(params)
+            Location location = locationService.createOrUpdateLocation(params)
             if (!location.validate() || !location.save(failOnError: true)) {
                 throw new ValidationException("Invalid location ${location.name}", location.errors)
             }
         }
-    }
-
-    Location createOrUpdateLocation(Map params) {
-        Location location
-        if (params.id) {
-            location = Location.findById(params.id)
-        } else if (params.locationNumber) {
-            location = Location.findByNameOrLocationNumber(params.name, params.locationNumber)
-        } else {
-            location = Location.findByName(params.name)
-        }
-        if (!location) {
-            location = new Location()
-        }
-
-        location.name = params.name
-        location.active = Boolean.valueOf(params.active)
-        location.locationNumber = params.locationNumber
-        location.locationGroup = params.locationGroup ? LocationGroup.findByName(params.locationGroup) : null
-        location.parentLocation = params.parentLocation ? Location.findByNameOrLocationNumber(params.parentLocation, params.parentLocation) : null
-        location.organization = params.organization ? Organization.findByCodeOrName(params.organization, params.organization) : null
-        location.address = createOrUpdateAddress(params, location?.address?.id)
-        if (!location.inventory && !location.parentLocation) {
-            location.inventory = new Inventory(['warehouse': location])
-        }
-        def locationType = null
-        if (params.locationType) {
-            String locationTypeName = LocalizationUtil.getDefaultString(params.locationType as String)
-            // TODO: Replace with a single GORM .find with Closure when in Grails 3 (available since Grails 2.0)
-            LocationType matchedLocationType = LocationType
-                    .findAllByNameLike(locationTypeName + "%")
-                    .find{ LocalizationUtil.getDefaultString(it.name) == locationTypeName}
-            locationType = matchedLocationType
-        }
-
-        def currentLocationType = location.locationType
-
-        //Do not allow to change internal type location to non-internal
-        if (((currentLocationType?.isInternalLocation() || currentLocationType?.isZone()) && (locationType?.isInternalLocation() || locationType?.isZone())) || !(currentLocationType?.isInternalLocation() || currentLocationType?.isZone())) {
-            location.locationType = locationType
-            location.supportedActivities = currentLocationType && locationType == currentLocationType ? location.supportedActivities : location.supportedActivities?.clear()
-        }
-
-        // Add required association to organization for depots and suppliers
-        if (!(location.locationType?.isInternalLocation() || location.locationType?.isZone()) && !location.organization) {
-            def locationCode = identifierService.generateOrganizationIdentifier(params.organization)
-            Organization organization = (location.locationType?.locationTypeCode == LocationTypeCode.SUPPLIER) ?
-                    organizationService.findOrCreateSupplierOrganization(params?.organization, locationCode) :
-                    organizationService.findOrCreateOrganization(params?.organization, locationCode)
-
-            location.organization = organization
-        }
-
-        return location
-    }
-
-    Address createOrUpdateAddress(Map params, @Nullable String addressId) {
-        Address address
-        if (addressId) {
-            address = Address.findById(addressId)
-        }
-
-        if (!address) {
-            address = new Address()
-        }
-
-        address.address = params.streetAddress ?: ''
-        address.address2 = params.streetAddress2 ?: ''
-        address.city = params.city ?: ''
-        address.stateOrProvince = params.stateOrProvince ?: ''
-        address.postalCode = params.postalCode ?: ''
-        address.country = params.country ?: ''
-        address.description = params.description ?: ''
-
-        if (address.validate() && !address.hasErrors()) {
-            address.save()
-        }
-
-        return address
     }
 }
